Scholership_finder/requirements.txt
pandas 
google-generativeai